requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio @ file:///Users/ec2-user/ci_py311/anyio_1678317412955/work/dist
appnope @ file:///Users/ec2-user/ci_py311/appnope_1678317440516/work
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///Users/ec2-user/ci_py311/argon2-cffi-bindings_1678317076583/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
attrs==23.1.0
Babel @ file:///Users/ec2-user/ci_py311/babel_1678318085010/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_bczxn9gqkq/croot/beautifulsoup4-split_1681493044307/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
brotlipy==0.7.0
certifi @ file:///Users/ec2-user/ci_py311/certifi_1678321191472/work/certifi
cffi @ file:///Users/ec2-user/ci_py311/cffi_1678315312775/work
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
comm @ file:///Users/ec2-user/ci_py311/comm_1678317779525/work
cryptography @ file:///Users/ec2-user/ci_py311_2/cryptography_1679335457662/work
debugpy @ file:///Users/ec2-user/ci_py311/debugpy_1678317801115/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints @ file:///Users/ec2-user/ci_py311/entrypoints_1678316169365/work
et-xmlfile==1.1.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///Users/ec2-user/ci_py311_2/python-fastjsonschema_1679338737111/work
Flask==2.2.3
flask-ngrok==0.0.25
frozenlist==1.3.3
future==0.18.3
idna @ file:///Users/ec2-user/ci_py311/idna_1678315819133/work
ipykernel @ file:///Users/ec2-user/ci_py311/ipykernel_1678318220696/work
ipython @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_cd4ia5pr17/croot/ipython_1680701875298/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==2.1.2
jedi @ file:///Users/ec2-user/ci_py311_2/jedi_1679336327335/work
Jinja2 @ file:///Users/ec2-user/ci_py311/jinja2_1678317138271/work
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_f9p98o1kqf/croot/jsonschema_1678983430226/work
jupyter-server @ file:///Users/ec2-user/ci_py311/jupyter_server_1678317903958/work
jupyter_client @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_5e4bbpqn9e/croot/jupyter_client_1680171866753/work
jupyter_core @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_b82fz_h369/croot/jupyter_core_1679906581737/work
jupyterlab @ file:///Users/ec2-user/ci_py311/jupyterlab_1678323572026/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab_server @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_9039tf6tpv/croot/jupyterlab_server_1680792539169/work
line-bot-sdk==2.4.2
lxml @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_cccmd60j4n/croot/lxml_1679646460589/work
MarkupSafe @ file:///Users/ec2-user/ci_py311/markupsafe_1678316973889/work
matplotlib-inline @ file:///Users/ec2-user/ci_py311/matplotlib-inline_1678317544666/work
mistune @ file:///Users/ec2-user/ci_py311/mistune_1678317272228/work
multidict==6.0.4
mwclient==0.10.1
nbclassic @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_e9be4uk5qu/croot/nbclassic_1681756175065/work
nbclient @ file:///Users/ec2-user/ci_py311/nbclient_1678317300721/work
nbconvert @ file:///Users/ec2-user/ci_py311/nbconvert_1678317567483/work
nbformat @ file:///Users/ec2-user/ci_py311/nbformat_1678317010390/work
nest-asyncio @ file:///Users/ec2-user/ci_py311/nest-asyncio_1678316288195/work
notebook @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_be_2kh1kd4/croot/notebook_1681756176866/work
notebook_shim @ file:///Users/ec2-user/ci_py311/notebook-shim_1678318293647/work
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.4
openpyxl==3.1.2
packaging @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_c3zlhpboab/croot/packaging_1678965318579/work
pandas==2.0.0
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
platformdirs @ file:///Users/ec2-user/ci_py311/platformdirs_1678315988118/work
prometheus-client @ file:///Users/ec2-user/ci_py311_2/prometheus_client_1679340232148/work
prompt-toolkit @ file:///Users/ec2-user/ci_py311/prompt-toolkit_1678317707969/work
psutil @ file:///Users/ec2-user/ci_py311_2/psutil_1679337242143/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_19gfn1ib_u/croot/pyopenssl_1678965300171/work
pyrsistent @ file:///Users/ec2-user/ci_py311/pyrsistent_1678316053523/work
PySocks @ file:///Users/ec2-user/ci_py311/pysocks_1678315868424/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///Users/ec2-user/ci_py311/pytz_1678318034461/work
pyzmq @ file:///Users/ec2-user/ci_py311/pyzmq_1678316754714/work
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
scipy==1.10.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///Users/ec2-user/ci_py311/sniffio_1678317345022/work
soupsieve @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_a42j8ggi70/croot/soupsieve_1680518482618/work
spicy==0.16.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
structlog==23.1.0
terminado @ file:///Users/ec2-user/ci_py311/terminado_1678317735543/work
tiktoken==0.3.3
tinycss2 @ file:///Users/ec2-user/ci_py311/tinycss2_1678317366076/work
tomli @ file:///Users/ec2-user/ci_py311/tomli_1678315243316/work
tornado @ file:///Users/ec2-user/ci_py311/tornado_1678316833426/work
tqdm==4.65.0
traitlets @ file:///Users/ec2-user/ci_py311/traitlets_1678316097905/work
typing_extensions @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_cazhnzmri4/croot/typing_extensions_1681939520610/work
tzdata==2023.3
urllib3 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_61c6ppfd7m/croot/urllib3_1680254700568/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///Users/ec2-user/ci_py311/websocket-client_1678317758506/work
Werkzeug==2.2.3
yarl==1.8.2
